Create to_s method override in Hangman
Marshal is used to create a deep copy of grid as most ways of doing this
for arrays work well for 1D but not for 2D. This is because doing
something like grid[0..] returns a deep copy of an array of pointers to
the same arrays as in the original. This means altering these arrays
will alter the original grid so the copy is not actually deep. It is
good practice to avoid mutating array arguments within methods.