Create Player with init and user_input methods
Basically I made the char input work, and the game is nearly playable.

I started on saving the game but decided I should focus more on
finishing the game proper first, so I shall ignore all methods related
to saving the game for testing such as serialize in Hangman and save in
Player.