Create play method in Game
The game now works as intended.

The MAX_MISTAKES constant was moved to Game to allow it to be used there
which means that it now has to be passed to Hangman as an init arg. I
think it is valid to interpret this as a game setting anyway.

The above constant change also meant initializing the objects directly
in the init method of Game rather than being passed in as args. This
will work just fine in this case, and simplifies the message sending
from main.rb anyway.