Create save functionality in Player
The only thing left is to serialize in Hangman.

I removed input as an instance variable in Player since the code reads
better with the multi-line loop and not having receive_input do too much
and be used weirdly in the loop. Now input is returned from the
user_input method instead of being read from the instance variable.