Create overall Game serializer
I realized that the instance of Executioner also needed to be saved, and
I then realized that the whole instance of Game needed to be saved too!
So I created serializer methods for each and chained them together in
Game.

I chained the serializers in this way to avoid Game needing to know more
about Hangman and Executioner than it needed to (i.e. the instance
variables). It would also allow all objects to be separately serialized
as needed.