Load in saved game at the start
Created deserialize methods in Hangman, Executioner, and Game to parse
the JSON save data and update the objects with the correct data.

I moved requiring json to main.rb since it only needs to be required in
one file so it makes sense that it is in the main file.